+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Include example directory that installs the library and uses it.
 - Added .npmignore file to ignore src/ directory.
 - Using babel-cli and some related modules to compile for npm usage.
+- Introducing MVW pattern
 
 ### Fixed
 - We can now use the library without having to have ES6 features.
@@ -16,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 - Changed package.json "main" to lib/go-patterns.js.
+- Modified babel presets and plugins
 
 ## [v1.0.2](https://github.com/pgarciacamou/go-patterns/releases/tag/v1.0.2) - 2017-04-23
 ### Added